% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dob_from_chi.R
\name{age_from_chi}
\alias{age_from_chi}
\title{Extract age from the CHI number}
\usage{
age_from_chi(
  chi_number,
  ref_date = NULL,
  min_age = 0,
  max_age = NULL,
  chi_check = TRUE
)
}
\arguments{
\item{chi_check}{}
}
\value{

}
\description{
\code{age_from_chi} takes a CHI number or a vector of CHI numbers
and returns the age as implied by the CHI number(s). If the DoB is ambiguous
it will return NA. It uses \code{dob_from_chi}.
}
