% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_lookup}
\alias{area_lookup}
\title{Codes and names of Scottish geographical and administrative areas.}
\format{A \code{\link[tibble]{tibble}} with 2 variables and over 23,000 rows:
\describe{
  \item{geo_code}{Standard geography code - 9 characters}
  \item{area_name}{Name of the area the code represents}
}}
\source{
\url{https://statistics.gov.scot/}
}
\usage{
area_lookup
}
\description{
A dataset containing the codes and names of areas used in Scotland.
}
\seealso{
The script used to create the area_lookup dataset:
\url{https://github.com/Health-SocialCare-Scotland/phsmethods/blob/geo_names/data-raw/area_lookup.R}
}
\keyword{datasets}
