% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_name_lookup}
\alias{area_name_lookup}
\title{Codes and names of Scottish geographical and administrative areas.}
\format{A data frame with 2 variables and over 23000 rows:
\describe{
  \item{geo_code}{Standard geography code - 9 characters}
  \item{area_name}{Name of the area the code represents}
}}
\source{
\url{https://statistics.gov.scot/}
}
\usage{
area_name_lookup
}
\description{
A dataset containing the codes and names of areas used in Scotland.
}
\keyword{datasets}
