% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.R
\name{chi_check}
\alias{chi_check}
\alias{chi_pad}
\title{chi functions}
\usage{
chi_check(to_check)

chi_pad(to_check)
}
\arguments{
\item{to_check}{\code{character} vector of CHI numbers}
}
\value{
character
}
\description{
The chi functions operate on CHI numbers:

\itemize{
\item \code{chi_pad} adds a leading zero to 9 character CHIs
\item \code{chi_check} tests a CHI for validity
}
}
\details{
\code{chi_check}:

This checks for a valid CHI by:

\itemize{
\item checking for invalid characters (non-numeric)
\item checking for length of 10 characters
\item checking that first 6 characters parse to valid date
\item checksum digit is correct
}

\code{chi_pad}:

Depending on the source, CHI numbers are sometimes mising a leading zero.
\code{chi_pad} takes a 9 digit CHI and adds a leading zero. Only CHI numbers
of 9 characters are changed.
}
\examples{
x <- c("0101011237", "0101201234","3201201234", "0113201234",
 "3213201234", "123456789", "12345678900", "010120123?")
chi_check(x)

#chi_pad differs from str_pad()
#as only 9 character CHIs are changed
x <- c("1234567890", "123456789", "123")
chi_pad(x)
}
