% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_pad.R
\name{chi_pad}
\alias{chi_pad}
\title{Add a leading zero to nine-digit CHI numbers}
\usage{
chi_pad(chi_number, chi_check = FALSE, check_mod11 = TRUE, check_mod10 = TRUE)
}
\arguments{
\item{chi_number}{a CHI number or a vector of CHI numbers with \code{character} class.}

\item{chi_check}{logical, optionally check the CHI for validity (after
padding), any invalid CHIs will be substituted with \code{NA}. The default (FALSE)
will skip checking but we recommend you check any padded CHIs.This uses
\code{\link[=chi_check]{chi_check()}}.}

\item{check_mod11, check_mod10}{Logical values (TRUE or FALSE, default is \code{TRUE}). By default, a CHI that passes either the modulo 10 or the modulo 11 check will be considered valid. Historically, CHIs only used modulo 11 for their check digit; however, starting in August 2026, some CHIs will only pass if they meet the modulo 10 criteria.
Implementation of Mod 10 CHI numbers is scheduled for August 2026.
From this date, CHI numbers are valid if they pass either a Mod 11 check
or a Mod 10 check.}
}
\value{
The original character vector with CHI numbers padded if applicable.
}
\description{
\code{chi_pad} takes a nine-digit CHI number with
\code{character} class and prefixes it with a zero. Any values provided
which are not a string comprised of nine numeric digits remain unchanged.
}
\details{
The Community Health Index (CHI) is a register of all patients in
NHS Scotland. A CHI number is a unique, ten-digit identifier assigned to
each patient on the index.

The first six digits of a CHI number are a patient's date of birth in
DD/MM/YY format. The first digit of a CHI number must, therefore, be 3 or
less. Depending on the source, CHI numbers are sometimes missing a leading
zero.

While a CHI number is made up exclusively of numeric digits, it cannot be
stored with \code{numeric} class in R. This is because leading zeros in
numeric values are silently dropped, a practice not exclusive to R. For this
reason, \code{chi_pad} accepts input values of \code{character} class
only, and returns values of the same class.
}
\examples{
chi_pad(c("101011237", "101201234"))
}
