% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_groups.R
\name{create_age_groups}
\alias{create_age_groups}
\title{Create age groups}
\usage{
create_age_groups(
  x,
  from = lifecycle::deprecated(),
  to = lifecycle::deprecated(),
  by = lifecycle::deprecated(),
  breaks = seq(0, 90, 5),
  as_factor = FALSE
)
}
\arguments{
\item{x}{a vector of numeric values}

\item{from}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{breaks} instead.}

\item{to}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{breaks} instead.}

\item{by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{breaks} instead.}

\item{breaks}{a numeric vector of cut points defining the age groups. The
default is \code{seq(0, 90, 5)}, which corresponds to the
\href{https://www.opendata.nhs.scot/dataset/standard-populations/resource/edee9731-daf7-4e0d-b525-e4c1469b8f69}{European Standard Population}
age groups.}

\item{as_factor}{The default behaviour is to return a character vector. Use
\code{TRUE} to return a factor vector instead.}
}
\value{
A character vector, where each element is the age group for the
corresponding element in \code{x}. If \code{as_factor = TRUE}, a factor
vector is returned instead.
}
\description{
\code{create_age_groups()} takes a numeric vector and assigns each age to the
appropriate age group.
}
\details{
The \code{breaks} vector defines the cut points for the age groups. The final
age group will capture all ages equal to or greater than the last value in
\code{breaks}, labelled as \verb{last+}. If the cut points are not evenly spaced, the
labels will reflect the actual width of each group.
}
\examples{
age <- c(54, 7, 77, 1, 26, 101)

create_age_groups(age)
create_age_groups(age, breaks = seq(0, 80, 10))

# Final group may start below the last break
create_age_groups(age, breaks = seq(0, 65, 10))

# To get the output as a factor:
create_age_groups(age, as_factor = TRUE)
}
