% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_size.R
\name{file_size}
\alias{file_size}
\title{Calculate file size}
\usage{
file_size(filepath = getwd(), pattern = NULL)
}
\arguments{
\item{filepath}{A filepath. Defaults to the working directory,
\code{getwd()}.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
A \code{\link[tibble]{tibble}} listing the names of the files within
\code{filepath} which match \code{pattern} and their respective file sizes.
If no \code{pattern} is specified, \code{file_size} returns the names and
file sizes of all files within \code{filepath}.

If \code{filepath} is an empty folder, or \code{pattern} matches no files
within \code{filepath}, \code{file_size} returns \code{NULL}.
}
\description{
\code{file_size} takes a filepath and an optional regex pattern.
It returns the size of any files within that directory which match the given
pattern.
}
\seealso{
For more information on using regular expressions, see this
\href{https://www.jumpingrivers.com/blog/regular-expressions-every-r-programmer-should-know/}{Jumping Rivers blog post}
and this
\href{https://stringr.tidyverse.org/articles/regular-expressions.html}{vignette}
from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
