% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_year.R
\name{fin_year}
\alias{fin_year}
\title{Assign a date to a financial year}
\usage{
fin_year(date)
}
\arguments{
\item{date}{A date which must be supplied with \code{Date} class. The
functions as.Date() or lubridate::dmy() are examples of functions that can
be used to change a variable to date class.}
}
\description{
\code{fin_year} takes a date and assigns it to the correct
financial year in the PHI specified format.
}
\details{
The PHI accepted format for financial year is yyyy/yy e.g. 2017/18.
}
\examples{
x <- lubridate::dmy(c("21012017", "04042017", "17112017"))

fin_year(x)

}
