% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_get_resource.R
\name{opendata_ds_dump_url}
\alias{opendata_ds_dump_url}
\title{Creates the URL for the datastore dump end-point}
\usage{
opendata_ds_dump_url(res_id)
}
\arguments{
\item{res_id}{a resource ID}
}
\value{
a url
}
\description{
Creates the URL for the datastore dump end-point
}
