% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_get_resource.R
\name{opendata_ds_search_url}
\alias{opendata_ds_search_url}
\title{Creates the URL for the datastore search end-point}
\usage{
opendata_ds_search_url()
}
\value{
a url
}
\description{
Creates the URL for the datastore search end-point
}
