% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_get_full_resource.R
\name{opendata_get_full_resource}
\alias{opendata_get_full_resource}
\title{Get full Open Data resource}
\usage{
opendata_get_full_resource(res_id)
}
\arguments{
\item{res_id}{The resource ID as found on \href{https://www.opendata.nhs.scot/}{NHS Open Data platform}}
}
\value{
a [tibble][tibble::tibble-package] with the data
}
\description{
Get full Open Data resource
}
\examples{
opendata_get_full_resource("d5ffffc0-f6f3-4b76-8f38-71ccfd7747a4")
}
