% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_get_resource.R
\name{opendata_get_resource}
\alias{opendata_get_resource}
\title{Get Open Data resource}
\usage{
opendata_get_resource(res_id, rows = NULL)
}
\arguments{
\item{res_id}{The resource ID as found on \href{https://www.opendata.nhs.scot/}{NHS Open Data platform}}

\item{rows}{(optional) specify the max number of rows to return
use this when testing code to reduce the size of the request
it will default to the either all the data or 99,999 (the hard limit on this API endpoint)}
}
\value{
a [tibble][tibble::tibble-package] with the data
}
\description{
Get Open Data resource
}
\examples{
opendata_get_resource(res_id = "a794d603-95ab-4309-8c92-b48970478c14")
}
