% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_get_dataset.R
\name{opendata_package_show_url}
\alias{opendata_package_show_url}
\title{Creates the URL for the package_show end-point}
\usage{
opendata_package_show_url()
}
\value{
a url
}
\description{
Creates the URL for the package_show end-point
}
