% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_get_resource.R
\name{opendata_ua}
\alias{opendata_ua}
\title{Open Data user agent}
\usage{
opendata_ua()
}
\value{
a {httr} user_agent string
}
\description{
This is used internally to return a standard useragent
Supplying a user agent means requests using the package
can be tracked more easily
}
