% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{as_percent}
\alias{as_percent}
\title{Percentages}
\usage{
as_percent(x)
}
\arguments{
\item{x}{\link{numeric} vector of proportions.}
}
\value{
A class of object \code{percent}.
}
\description{
\code{percent} is a lightweight S3 class allowing for pretty
printing of proportions as percentages.
}
\details{
By default all percentages are formatted to 2 decimal places which can be
overwritten using \code{format()}. It's worth noting that the digits argument in
\code{format.percent} uses decimal rounding instead of significant digit rounding.
}
\examples{
library(percent)

# Convert proportions to percentages
as_percent(seq(0, 1, 0.1))

# You can use round() as usual
p <- as_percent(15.56 / 100)
round(p)
round(p, digits = 1)

p2 <- as_percent(0.0005)
signif(p2, 2)
floor(p2)
ceiling(p2)

# We can do basic math operations as usual

# Order of operations matters
10 * as_percent(c(0, 0.5, 2))
as_percent(c(0, 0.5, 2)) * 10

as_percent(0.1) + as_percent(0.2)

# Formatting options
format(as_percent(2.674 / 100), digits = 2, symbol = " (\%)")
# Prints nicely in data frames (and tibbles)
library(dplyr)
starwars \%>\%
  count(eye_color) \%>\%
  mutate(perc = as_percent(n/sum(n))) \%>\%
  arrange(desc(perc))  \%>\% # We can do numeric sorting with percent vectors
  mutate(perc_rounded = round(perc))
}
