% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phimethods.R
\docType{package}
\name{phimethods}
\alias{phimethods}
\alias{phimethods-package}
\title{\code{phimethods} package}
\description{
Standard Methods for use in PHI.
}
\details{
See the README on \href{https://github.com/Health-SocialCare-Scotland/phimethods#readme}{GitHub}.
}
