% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsmethods.R
\docType{package}
\name{phsmethods}
\alias{phsmethods}
\title{\code{phsmethods} package}
\description{
Standard Methods for use in PHS.
}
\details{
See the README on \href{https://github.com/Health-SocialCare-Scotland/phsmethods#readme}{GitHub}.
}
