% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr.R
\name{qtr_year}
\alias{qtr_year}
\title{Find a quarter to a date}
\usage{
qtr_year(date, format = c("long", "short"))
}
\arguments{
\item{date}{A date supplied with \code{Date} class.}

\item{format}{A \code{character} string specifying the format the quarter
should be displayed in. Valid options are `short` (Jan-Mar 2018) and `long`
(January to March 2018).}
}
\description{
The qtr functions take a date input and calculate the relevant
quarter-related value from it.

\itemize{
\item \code{qtr_year} returns the current quarter and year in either short or
long format.

\item \code{qtr_end} returns the last month in the quarter in either short or
long format.

\item \code{qtr_prev} returns the previous quarter in either short or long
format.
}
}
\examples{
x <- lubridate::dmy(c("26032012", "04052012", "23092012"))

qtr_year(x, format = "long")

qtr_end(x, format = "short")

qtr_prev(x, format = "long")
}
