% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{postcode}
\alias{age_group}
\alias{fin_year}
\title{Functions renamed in December 2021}
\usage{
postcode(x, format = c("pc7", "pc8"))

age_group(x, from = 0, to = 90, by = 5, as_factor = FALSE)

fin_year(date)
}
\description{
`r lifecycle::badge('deprecated')`

We renamed a number of function to improve code clarity.
The old functions are still usable but will produce a warning.
After a reasonable amount of time they will be removed completely.

* `postcode()` -> `format_postcode()`
* `age_group()` -> `create_age_groups()`
* `fin_year()` -> `extract_fin_year()`
}
\keyword{internal}
