% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dob_from_chi.R
\name{dob_from_chi}
\alias{dob_from_chi}
\title{Extract Date of Birth (DoB) from the CHI number}
\usage{
dob_from_chi(
  chi_number,
  min_date = NULL,
  max_date = NULL,
  min_age = 0,
  max_age = 120,
  chi_check = TRUE
)
}
\arguments{
\item{chi_check}{}
}
\value{

}
\description{
\code{dob_from_chi} takes a CHI number or a vector of CHI numbers
and returns the DoB as implied by the CHI number(s). If the DoB is ambiguous
it will return NA
}
