% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inheritance_checks.R
\name{make_inheritance_checks}
\alias{make_inheritance_checks}
\title{Performs checks if elements have classes specified by 'target_classes' and prints out each element which is lacking the specified class or is NULL.
Throws an exception if one element does not pass the checks.
Element must have AT LEAST ONE of the specified classes to pass.}
\usage{
make_inheritance_checks(elements, target_classes)
}
\arguments{
\item{elements}{a list object containing element_name=element pairs for each element. Element names must be provided or else they will be ignored.}

\item{target_classes}{array of classes (string).}
}
\description{
Performs checks if elements have classes specified by 'target_classes' and prints out each element which is lacking the specified class or is NULL.
Throws an exception if one element does not pass the checks.
Element must have AT LEAST ONE of the specified classes to pass.
}
