% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{postcode}
\title{Functions renamed in December 2021}
\usage{
postcode(x, format = c("pc7", "pc8"))
}
\description{
`r lifecycle::badge('deprecated')`

We renamed a number of function to improve code clarity.
The old functions are still usable but will produce a warning.
After a reasonable amount of time they will be removed completely.

* `postcode()` -> `format_postcode()`
}
\keyword{internal}
