% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{postcode}
\alias{age_group}
\alias{fin_year}
\title{Functions renamed in phsmethods 0.2.1 (Dec 2021)}
\usage{
postcode(x, format = c("pc7", "pc8"))

age_group(x, from = 0, to = 90, by = 5, as_factor = FALSE)

fin_year(date)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

phsmethods 0.2.1 renamed a number of functions
to improve code clarity. The old functions will still
work but will produce a warning. After a reasonable
amount of time they will be removed completely.
\itemize{
\item \code{postcode()} -> \code{format_postcode()}
\item \code{age_group()} -> \code{create_age_groups()}
\item \code{fin_year()} -> \code{extract_fin_year()}
}
}
\keyword{internal}
